create empty data dir
Without the data/<ServiceBindings|ServiceInstances>.json files the broker will not work so adding defaults. Also, since it makes little sense to have two paths for data JSON files, we are moving the assets/catalog.json to data and fixed the default assets/config.json to point to the data dir.